Add default AcknowPackageParser parse method
