+++ Examples/AcknowExampleSPM/AcknowExampleSPM.xcodeproj/xcshareddata/xcschemes/AcknowExampleSPM.xcscheme
@@ -28,6 +28,16 @@
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
+         <TestableReference
+            skipped = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "AcknowListTests"
+               BuildableName = "AcknowListTests"
+               BlueprintName = "AcknowListTests"
+               ReferencedContainer = "container:../..">
+            </BuildableReference>
+         </TestableReference>
       </Testables>
    </TestAction>
    <LaunchAction
