+++ Sources/AcknowList/AcknowPodParser.swift
@@ -132,22 +132,4 @@ open class AcknowPodParser {
         let singleNewLineFinder = try! NSRegularExpression(pattern: "(?<=.)(\\h)*(\\R)(\\h)*(?=.)")
         return singleNewLineFinder.stringByReplacingMatches(in: text, range: NSRange(0..<text.count), withTemplate: " ")
     }
-
-    /**
-     Finds the first link (URL) in a given string.
-
-     @param text The string to parse.
-
-     @return The first link found, or `nil` if no link can be found.
-     */
-    class func firstLink(in text: String) -> URL? {
-        let types: NSTextCheckingResult.CheckingType = [.link]
-
-        guard let linkDetector = try? NSDataDetector(types: types.rawValue),
-            let firstLink = linkDetector.firstMatch(in: text, options: [], range: NSMakeRange(0, text.count)) else {
-                return nil
-        }
-
-        return firstLink.url
-    }
 }
